requirements.txt
boto3==1.26.87
botocore==1.29.87
cffi==1.15.1
ecdsa==0.18.0
greenlet==2.0.2
install==1.3.5
jmespath==1.0.1
mysql-connector-python==8.0.32
protobuf==3.20.3
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
rsa==4.9
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==2.0.5.post1
typing_extensions==4.5.0
urllib3==1.26.14
python-dateutil==2.8.2
